Refatoração de rotas e controladores na API
- `AccountController`: rota alterada para `api/accounts`; método `SignUp` adicionado (ex-`SignUpController`); tratamento de exceções implementado; rota `GetAccount` ajustada.
- `DepositController` e `OrderController`: rotas padronizadas para plurais (`api/deposits`, `api/orders`); métodos renomeados para refletir ações (`placeDeposit`, `placeOrder`).
- `SignUpController`: removido; lógica migrada para `AccountController`.
- Testes: atualizados para refletir novas rotas e cenários de erro; testes do `SignUpController` movidos para `AccountControllerTests`.
- Objetivo: consolidar funcionalidades, padronizar rotas RESTful e melhorar validação.
