+++ API/Controllers/AccountController.cs
@@ -1,4 +1,5 @@
 ﻿using API.DTOs;
+using Application.DTOs;
 using Application.Ports.Driving;
 using Domain.Exceptions;
 using Microsoft.AspNetCore.Mvc;
@@ -6,11 +7,39 @@ using System.Net;
 
 namespace API.Controllers
 {
-    [Route("api/account")]
+    [Route("api/accounts")]
     [ApiController]
     public class AccountController : ControllerBase
     {
-        [HttpGet]
+        [HttpPost("signUp")]
+        public async Task<IActionResult> SignUp(
+            [FromServices] ISignUp _signUpUseCase,
+            [FromBody] SignUpDto signUpDto)
+        {
+            try
+            {
+                var accountId = await _signUpUseCase.SingUpAsync(signUpDto);
+                return Ok(accountId);
+            }
+            catch (ValidationException ex)
+            {
+                return StatusCode(422, new ErrorResponseDto
+                {
+                    ErrorMessages = ex.Errors
+                        .Select(x => x.ErrorMessage)
+                        .ToList()
+                });
+            }
+            catch (Exception ex)
+            {
+                return StatusCode((int)HttpStatusCode.InternalServerError, new ErrorResponseDto
+                {
+                    ErrorMessages = new List<string> { ex.Message }
+                });
+            }
+        }
+
+        [HttpGet("getAccount")]
         public async Task<IActionResult> GetAccount(
             [FromServices] IGetAccount _getAccount,
             [FromQuery] Guid accountId)
